Improve unwind info code organisation
Split it in different files, added some more docs, and call the unwind
info optimisers directly.

Error handling improvements, more docs, and especially more tests coming
later. As this is a rather load-bearing part of the project, any issues
pop up pretty quickly in CI.

Test Plan
=========

CI
