+++ src/unwind_info/types.rs
@@ -1,0 +1,81 @@
+use lazy_static::lazy_static;
+
+#[repr(u8)]
+pub enum CfaType {
+    // Unknown = 0,
+    FramePointerOffset = 1,
+    StackPointerOffset = 2,
+    Expression = 3,
+    EndFdeMarker = 4,
+    UnsupportedRegisterOffset = 5,
+    OffsetDidNotFit = 6,
+}
+
+#[repr(u8)]
+pub enum RbpType {
+    // Unknown = 0,
+    CfaOffset = 1,
+    Register = 2,
+    Expression = 3,
+    UndefinedReturnAddress = 4,
+    OffsetDidNotFit = 5,
+}
+
+#[repr(u16)]
+pub enum PltType {
+    // Unknown = 0,
+    Plt1 = 1,
+    Plt2 = 2,
+}
+
+#[derive(Debug, Default, Copy, Clone)]
+pub struct CompactUnwindRow {
+    pub pc: u64,
+    pub cfa_type: u8,
+    pub rbp_type: u8,
+    pub cfa_offset: u16,
+    pub rbp_offset: i16,
+}
+
+impl CompactUnwindRow {
+    pub fn end_of_function_marker(last_addr: u64) -> CompactUnwindRow {
+        CompactUnwindRow {
+            pc: last_addr,
+            cfa_type: CfaType::EndFdeMarker as u8,
+            ..Default::default()
+        }
+    }
+}
+
+lazy_static! {
+    pub static ref PLT1: [u8; 11] = [
+        gimli::constants::DW_OP_breg7,
+        gimli::constants::DW_OP_const1u,
+        gimli::constants::DW_OP_breg16,
+        gimli::DwOp(0), // ?
+        gimli::constants::DW_OP_lit15,
+        gimli::constants::DW_OP_and,
+        gimli::constants::DW_OP_lit11,
+        gimli::constants::DW_OP_ge,
+        gimli::constants::DW_OP_lit3,
+        gimli::constants::DW_OP_shl,
+        gimli::constants::DW_OP_plus,
+    ].map(|a| a.0);
+
+    pub static ref PLT2: [u8; 11] = [
+        gimli::constants::DW_OP_breg7,
+        gimli::constants::DW_OP_const1u,
+        gimli::constants::DW_OP_breg16,
+        gimli::DwOp(0), // ?
+        gimli::constants::DW_OP_lit15,
+        gimli::constants::DW_OP_and,
+        gimli::constants::DW_OP_lit10,
+        gimli::constants::DW_OP_ge,
+        gimli::constants::DW_OP_lit3,
+        gimli::constants::DW_OP_shl,
+        gimli::constants::DW_OP_plus,
+    ].map(|a| a.0);
+}
+
+pub const RBP_X86: gimli::Register = gimli::Register(6);
+pub const RSP_X86: gimli::Register = gimli::Register(7);
